🚚 Rename: KakaoInit을 shared/providers로 이동
